данные запроса приходят, осталось правильно сформировать заказ и сохранить в бд

+++ src/main/java/com/koshelev/spring/web/entities/Product.java
@@ -22,6 +22,6 @@ public class Product {
     private String title;
 
     @Column(name = "cost")
-    private Double cost;
+    private Integer cost;
 
 }
